requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
requests==2.31.0
email-validator==2.1.0
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.9
SQLAlchemy==2.0.41
greenlet==3.0.3
psycopg2-binary==2.9.9
alembic==1.13.2
